+++ x/genutil/client/legacy/v0_18/migrate_test.go
@@ -1,10 +1,10 @@
-package v017
+package v018
 
 import (
 	"github.com/cosmos/cosmos-sdk/codec"
 	"github.com/cosmos/cosmos-sdk/x/genutil"
-	v017evm "github.com/okex/exchain/x/evm/legacy/v0_17"
-	v017staking "github.com/okex/exchain/x/staking/legacy/v0_17"
+	v018evm "github.com/okex/exchain/x/evm/legacy/v0_18"
+	v018staking "github.com/okex/exchain/x/staking/legacy/v0_18"
 	"github.com/stretchr/testify/require"
 	"testing"
 )
@@ -14,8 +14,8 @@ func TestMigrate(t *testing.T) {
 	v016Codec := codec.New()
 	codec.RegisterCrypto(v016Codec)
 
-	v017Codec := codec.New()
-	codec.RegisterCrypto(v017Codec)
+	v018Codec := codec.New()
+	codec.RegisterCrypto(v018Codec)
 
 	appState := genutil.AppMap{
 		"staking": []byte(`{"params":{"bond_denom":"okt","epoch":252,"max_bonded_validators":21,"max_validators_to_add_shares":30,"min_delegation":"0.000100000000000000","min_self_delegation":"10000.000000000000000000","unbonding_time":"1209600000000000"}}`),
@@ -24,13 +24,13 @@ func TestMigrate(t *testing.T) {
 	statsMigrate := Migrate(appState)
 
 	// stakingState
-	var stakingState v017staking.GenesisState
-	v017Codec.MustUnmarshalJSON(statsMigrate[v017staking.ModuleName], &stakingState)
+	var stakingState v018staking.GenesisState
+	v018Codec.MustUnmarshalJSON(statsMigrate[v018staking.ModuleName], &stakingState)
 	require.NotNil(t, stakingState)
 
 	// evmState
-	var evmState v017evm.GenesisState
-	v017Codec.MustUnmarshalJSON(statsMigrate[v017evm.ModuleName], &evmState)
+	var evmState v018evm.GenesisState
+	v018Codec.MustUnmarshalJSON(statsMigrate[v018evm.ModuleName], &evmState)
 	require.True(t, evmState.Params.EnableContractDeploymentWhitelist)
 	require.True(t, evmState.Params.EnableContractBlockedList)
 }
