+++ x/staking/legacy/v0_18/migrate.go
@@ -1,4 +1,4 @@
-package v0_17
+package v0_18
 
 import "github.com/okex/exchain/x/staking/legacy/v0_11"
